Migrate unit tests for isArray
