+++ test/unit.html
@@ -34,7 +34,6 @@
   <script src="flatten_test_old.js"></script>
   <script src="get_test_old.js"></script>
   <script src="hasOwn_test_old.js"></script>
-  <script src="isArray_test_old.js"></script>
   <script src="isPlainObject_test_old.js"></script>
   <script src="merge_test_old.js"></script>
   <script src="processCommand_test_old.js"></script>
